# internal/api/openapi/openapi.yaml
openapi: "3.0.3"
info:
  title: Updater Service API
  version: "1.0.0"
  description: |
    RESTful API for the updater service. Desktop applications query this API to check for
    software updates, retrieve release information, and manage applications.

    All request and response bodies use `application/json`. Timestamps follow RFC 3339 format
    and versions follow semantic versioning (semver).

    ## Authentication

    Protected endpoints require a Bearer token in the `Authorization` header:

    ```
    Authorization: Bearer <api-key>
    ```

    The permission hierarchy is cumulative:

    | Level  | Includes           | Description                                       |
    |--------|--------------------|---------------------------------------------------|
    | read   | read               | Query releases, list applications                 |
    | write  | read, write        | Register releases, create applications            |
    | admin  | read, write, admin | Update and delete applications, delete releases   |

    Public endpoints (update checks, latest version, health, OpenAPI spec) do not require
    authentication. When authentication is disabled in the server configuration, all endpoints
    are accessible without a token.
servers:
  - url: /api/v1
    description: Versioned API prefix
  - url: /
    description: Root prefix (used by /health)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        API key authentication. Keys are configured on the server with permission levels.
        Permission hierarchy: read < write < admin.
  parameters:
    AppIdPath:
      name: app_id
      in: path
      required: true
      schema:
        type: string
      description: Application identifier
...
  schemas:
    Platform:
      type: string
      enum: [windows, linux, darwin, android, ios]
      description: Target operating system
    Architecture:
      type: string
      enum: [amd64, arm64, "386", arm]
      description: Target CPU architecture
...
    SortOrder:
      type: string
      enum: [asc, desc]
      description: Sort direction
...
  responses:
    BadRequest:
      description: Malformed or invalid request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error: bad_request
            message: Invalid JSON body
            code: INVALID_REQUEST
            timestamp: "2026-02-16T10:00:00Z"
...
security:
  - bearerAuth: []
paths:
  /updates/{app_id}/check:
    get:
      tags: [updates]
      summary: Check for updates (GET)
      description: |
        Check whether a newer version is available for a specific application, platform,
        and architecture. Parameters are provided as query strings.
      operationId: checkForUpdatesGet
      security: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
        - name: current_version
          in: query
          required: true
          schema:
            type: string
          description: Client's current semantic version
          example: "2.0.0"
        - name: platform
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Platform"
        - name: architecture
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Architecture"
        - name: allow_prerelease
          in: query
          schema:
            type: boolean
            default: false
          description: Include pre-release versions
        - name: include_metadata
          in: query
          schema:
            type: boolean
            default: false
          description: Include release metadata in response
      responses:
        "200":
          description: Update check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCheckResponse"
              examples:
                update_available:
                  summary: Update available
                  value:
                    update_available: true
                    latest_version: "2.1.0"
                    current_version: "2.0.0"
                    download_url: https://releases.example.com/app/2.1.0/app-windows-amd64.exe
                    checksum: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
                    checksum_type: sha256
                    file_size: 15728640
                    release_notes: Performance improvements and bug fixes
                    release_date: "2026-02-10T12:00:00Z"
                    required: false
                    minimum_version: "1.0.0"
                no_update:
                  summary: No update available
                  value:
                    update_available: false
                    current_version: "2.1.0"
                    required: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /check:
    post:
      tags: [updates]
      summary: Check for updates (POST)
      description: |
        Alternative endpoint that accepts update check parameters as a JSON body instead
        of query parameters. Useful for clients that prefer structured request bodies.
      operationId: checkForUpdatesPost
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCheckRequest"
            example:
              application_id: my-app
              current_version: "2.0.0"
              platform: windows
              architecture: amd64
              allow_prerelease: false
              include_metadata: false
      responses:
        "200":
          description: Update check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateCheckResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /updates/{app_id}/latest:
    get:
      tags: [updates]
      summary: Get latest version (path-based)
      description: |
        Retrieve the latest release for a given application, platform, and architecture.
        The application ID is provided as a path parameter.
      operationId: getLatestVersionPath
      security: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
        - name: platform
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Platform"
        - name: architecture
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Architecture"
        - name: allow_prerelease
          in: query
          schema:
            type: boolean
            default: false
          description: Include pre-release versions
        - name: include_metadata
          in: query
          schema:
            type: boolean
            default: false
          description: Include release metadata in response
      responses:
        "200":
          description: Latest release information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestVersionResponse"
              example:
                version: "2.1.0"
                download_url: https://releases.example.com/app/2.1.0/app-windows-amd64.exe
                checksum: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
                checksum_type: sha256
                file_size: 15728640
                release_notes: Performance improvements and bug fixes
                release_date: "2026-02-10T12:00:00Z"
                required: false
                metadata: {}
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /latest:
    get:
      tags: [updates]
      summary: Get latest version (query-based)
      description: |
        Compatibility alias for the latest version endpoint. The application ID is provided
        as a query parameter instead of a path segment.
      operationId: getLatestVersionQuery
      security: []
      parameters:
        - name: app_id
          in: query
          required: true
          schema:
            type: string
          description: Application identifier
          example: my-app
        - name: platform
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Platform"
        - name: architecture
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/Architecture"
        - name: allow_prerelease
          in: query
          schema:
            type: boolean
            default: false
          description: Include pre-release versions
        - name: include_metadata
          in: query
          schema:
            type: boolean
            default: false
          description: Include release metadata in response
      responses:
        "200":
          description: Latest release information
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LatestVersionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /updates/{app_id}/releases:
    get:
      tags: [releases]
      summary: List releases
      description: |
        List releases for an application with optional filtering, sorting, and pagination.
        Requires `read` permission.
      operationId: listReleases
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
        - name: platform
          in: query
          schema:
            $ref: "#/components/schemas/Platform"
          description: Filter by platform
        - name: architecture
          in: query
          schema:
            $ref: "#/components/schemas/Architecture"
          description: Filter by architecture
        - name: version
          in: query
          schema:
            type: string
          description: Filter by exact version
        - name: required
          in: query
          schema:
            type: boolean
          description: Filter by required flag
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
          description: Page size
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page offset
        - name: sort_by
          in: query
          schema:
            $ref: "#/components/schemas/SortBy"
          description: Sort field (default release_date)
        - name: sort_order
          in: query
          schema:
            $ref: "#/components/schemas/SortOrder"
          description: Sort direction (default desc)
      responses:
        "200":
          description: Paginated list of releases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReleasesResponse"
              example:
                releases:
                  - id: rel-001
                    version: "2.1.0"
                    platform: windows
                    architecture: amd64
                    download_url: https://releases.example.com/app/2.1.0/app-windows-amd64.exe
                    checksum: e3b0c44298fc1c149afbf4c8996fb924
                    checksum_type: sha256
                    file_size: 15728640
                    release_notes: Performance improvements
                    release_date: "2026-02-10T12:00:00Z"
                    required: false
                    minimum_version: "1.0.0"
                total_count: 1
                page: 1
                page_size: 50
                has_more: false
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /updates/{app_id}/register:
    post:
      tags: [releases]
      summary: Register release
      description: |
        Register a new release for an application. Requires `write` permission.
      operationId: registerRelease
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterReleaseRequest"
            example:
              application_id: my-app
              version: "2.1.0"
              platform: windows
              architecture: amd64
              download_url: https://releases.example.com/app/2.1.0/app-windows-amd64.exe
              checksum: e3b0c44298fc1c149afbf4c8996fb92427ae41e4649b934ca495991b7852b855
              checksum_type: sha256
              file_size: 15728640
              release_notes: Performance improvements and bug fixes
              required: false
              minimum_version: "1.0.0"
              metadata:
                build_number: "1234"
                commit_sha: abc123
      responses:
        "201":
          description: Release registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterReleaseResponse"
              example:
                id: rel-001
                message: Release registered successfully
                created_at: "2026-02-16T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /updates/{app_id}/releases/{version}/{platform}/{arch}:
    delete:
      tags: [releases]
      summary: Delete release
      description: |
        Delete a specific release identified by application, version, platform, and
        architecture. Requires `admin` permission.
      operationId: deleteRelease
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
        - $ref: "#/components/parameters/VersionPath"
        - $ref: "#/components/parameters/PlatformPath"
        - $ref: "#/components/parameters/ArchPath"
      responses:
        "200":
          description: Release deleted successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteReleaseResponse"
              example:
                id: rel-001
                message: Release deleted successfully
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /applications:
    get:
      tags: [applications]
      summary: List applications
      description: |
        Retrieve a paginated list of all registered applications. Requires `read` permission.
      operationId: listApplications
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            default: 50
          description: Page size
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Page offset
      responses:
        "200":
          description: Paginated list of applications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListApplicationsResponse"
              example:
                applications:
                  - id: my-app
                    name: My Application
                    description: A desktop application
                    platforms: [windows, linux, darwin]
                    created_at: "2026-01-01T00:00:00Z"
                    updated_at: "2026-02-01T00:00:00Z"
                total_count: 1
                page: 1
                page_size: 50
                has_more: false
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"

    post:
      tags: [applications]
      summary: Create application
      description: |
        Register a new application in the update service. Requires `write` permission.
      operationId: createApplication
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicationRequest"
            example:
              id: my-app
              name: My Application
              description: A cross-platform desktop application
              platforms: [windows, linux, darwin]
              config:
                auto_update: false
                update_interval: 3600
                required_update: false
                allow_prerelease: false
                analytics_enabled: false
      responses:
        "201":
          description: Application created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateApplicationResponse"
              example:
                id: my-app
                message: Application created successfully
                created_at: "2026-02-16T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"
  /applications/{app_id}:
    get:
      tags: [applications]
      summary: Get application
      description: |
        Retrieve detailed information about a single application, including configuration
        and statistics. Requires `read` permission.
      operationId: getApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
      responses:
        "200":
          description: Application details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationInfoResponse"
              example:
                id: my-app
                name: My Application
                description: A desktop application
                platforms: [windows, linux, darwin]
                config:
                  auto_update: false
                  update_interval: 3600
                  required_update: false
                  allow_prerelease: false
                  analytics_enabled: false
                stats:
                  total_releases: 12
                  latest_version: "2.1.0"
                  latest_release_date: "2026-02-10T12:00:00Z"
                  platform_count: 3
                  required_releases: 1
                created_at: "2026-01-01T00:00:00Z"
                updated_at: "2026-02-01T00:00:00Z"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"

    put:
      tags: [applications]
      summary: Update application
      description: |
        Update an existing application's metadata and configuration. Only provided fields
        are updated; omitted fields remain unchanged. Requires `admin` permission.
      operationId: updateApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicationRequest"
            example:
              name: My Application (Renamed)
              description: Updated description
              platforms: [windows, linux, darwin, android]
              config:
                auto_update: true
                update_interval: 1800
      responses:
        "200":
          description: Application updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateApplicationResponse"
              example:
                id: my-app
                message: Application updated successfully
                updated_at: "2026-02-16T11:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalError"

    delete:
      tags: [applications]
      summary: Delete application
      description: |
        Permanently delete an application. The application must have no existing releases;
        if releases exist the request returns `409 Conflict`. Requires `admin` permission.
      operationId: deleteApplication
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/AppIdPath"
      responses:
        "204":
          description: Application deleted successfully (no body)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /admin/keys:
    get:
      tags: [keys]
      summary: List API keys
      description: Returns metadata for all API keys. Requires admin permission.
      operationId: listAPIKeys
      security:
        - bearerAuth: []
      responses:
        "200":
          description: List of API key metadata
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIKeyMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [keys]
      summary: Create API key
      description: |
        Creates a new API key. The raw key is returned exactly once in the response body.
        Store it securely — it cannot be retrieved again.

        Requires admin permission.
      operationId: createAPIKey
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAPIKeyRequest"
      responses:
        "201":
          description: Key created. Raw key included once — store it now.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateAPIKeyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "500":
          $ref: "#/components/responses/InternalError"
...